[#11067] Split built-in IdP user mapper tests by backend
